+++ array.h
@@ -1,13 +1,14 @@
 #pragma once
 #ifndef ARRAY_H
 #define ARRAY_H
+#include <stdlib.h>
 
 struct Array;
 typedef struct Array Array;
 typedef void(*ArrayFunc)(void*);
 typedef int(*ArrayCriteria)(void*);
 
-Array* ArrayCreate(unsigned int size);
+Array* ArrayCreate(size_t size);
 void ArrayClear(Array *array, ArrayFunc clean);
 void ArrayAnnihilate(Array **array, ArrayFunc clean);
 #define ArrayPurge(array) ArrayAnnihilate(array, free)
@@ -15,14 +16,14 @@ void ArrayAnnihilate(Array **array, ArrayFunc clean);
 #define ArrayNew() ArrayCreate(1)
 #define ArrayClean(array) ArrayClear(array, free)
 
-unsigned int ArrayLength(Array *array);
-unsigned int ArraySize(Array *array);
-void ArrayInsert(Array *array, unsigned int index, void *data);
+size_t ArrayLength(Array *array);
+size_t ArraySize(Array *array);
+void ArrayInsert(Array *array, size_t index, void *data);
 void ArrayAppend(Array *array, void *data);
-void ArrayDeleteRange(Array *array, unsigned int start, unsigned int end);
-void **ArrayRemove(Array *array, unsigned int index);
+void ArrayDeleteRange(Array *array, size_t start, size_t end);
+void **ArrayRemove(Array *array, size_t index);
 void *ArrayPop(Array *array);
-void **ArrayRemoveRange(Array *array, unsigned int start, unsigned int end);
-void *ArrayElement(Array *array, unsigned int index);
+void **ArrayRemoveRange(Array *array, size_t start, size_t end);
+void *ArrayElement(Array *array, size_t index);
 void ArrayIterate(Array *array, ArrayFunc func);
 #endif // ARRAY_H
